create new class for marker
