+++ Assets/MixedRealityToolkit.Extensions/MarkerDetection/Marker.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c6000d06d83d25d4cad0e41778439161
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
